Dynamic Content replace slot with HTML
When a slot is replaced on a landingpages it uses the appendXML function. The the name suggests it requires (valid) XML. HTML certainly related to XML, but has some differences. For example, if and `&` is used the XML will get stressed.

No I had 2 cases where the replacement of the slot through an error:
- In the dynamic content there was een URL. The URL got replaced my mautic to enable tracking. But the mautic url (the token) contained a "&". Result: replacing the child failed.
- In my template I used an `&nbsp;`. This is a typical html code. It make the XML parse fail.

Just by wrapping the new Slot content into `<![CDATA[ **CONTENT**  ]]>` fixes this issue. It even supports invalid HTML.