Reject page_hit queue messages without a Hit
In consuming a page_hit queue, it is possible to encounter a message
which does not contain a valid 'hitId' (e.g. NULL). Since the
signature of the PageModel#processPageHit method requires the $hit
argument to be an instance of Hit, an exception will be thrown when no
entity is found by the HitRepository.

A null value for $hit is handled herein by simply setting the event to
REJECT that message. This will result in a loss of data, which may or
may not be a problem for some users. Should retention of that data be
desired, one should probably register an event listener of their own
to do something with rejected message payloads.